+++ test/tests/integration/server/Services/MyService.lua
@@ -1,6 +1,7 @@
 local Knit = require(game:GetService("ReplicatedStorage").Knit)
 local Option = require(Knit.Util.Option)
 local Timer = require(Knit.Util.Timer)
+local TableUtil = require(Knit.Util.TableUtil)
 
 
 local MyService = Knit.CreateService {
@@ -26,15 +27,38 @@ end
 
 
 function MyService:KnitStart()
+
 	print(self.Name .. " started")
+
 	local timer = Timer.new(1)
 	timer.Tick:Connect(function()
 		print("TICK", time())
 	end)
 	timer:Start()
+
+	local conn = Timer.Simple(1, function()
+		print("TICK_SIMPLE", time())
+	end)
+
+	local timerNoDrift = Timer.new(1)
+	timerNoDrift.AllowDrift = false
+	timerNoDrift.Tick:Connect(function()
+		print("TICK_SYNC", time())
+	end)
+	timerNoDrift:Start()
+
 	task.delay(5, function()
 		timer:Destroy()
+		timerNoDrift:Destroy()
+		conn:Disconnect()
 	end)
+
+	local a = {10, 20, 30, 40, 50}
+	local b = {5, 4, 3, 2, 1}
+	for i,v in TableUtil.Zip(a, b) do
+		print(i, v)
+	end
+
 end